+++ lib/kitchen/element_enumerator_base.rb
@@ -218,11 +218,19 @@ module Kitchen
                except: except)
     end
 
-    # Returns an enumerator that iterates through exercises within the scope of this enumerator
+    # Returns an enumerator that iterates through references within the scope of this enumerator
     #
     # @param css_or_xpath [String] additional selectors to further narrow the element iterated over;
     #   a "$" in this argument will be replaced with the default selector for the element being
     #   iterated over.
+    #
+    def references(css_or_xpath=nil)
+      block_error_if(block_given?)
+      chain_to(ReferenceElementEnumerator, css_or_xpath: css_or_xpath)
+    end
+
+    # Returns an enumerator that iterates through exercises within the scope of this enumerator
+    #
     # @param only [Symbol, Callable] the name of a method to call on an element or a
     #   lambda or proc that accepts an element; elements will only be included in the
     #   search results if the method or callable returns true