+++ lib/kitchen/references_element_enumerator.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module Kitchen
+  # An enumerator for table elements
+  #
+  class ReferenceElementEnumerator < ElementEnumeratorBase
+    # Returns a factory for this enumerator
+    #
+    # @return [ElementEnumeratorFactory]
+    #
+    def self.factory
+      ElementEnumeratorFactory.new(
+        default_css_or_xpath: '.reference', # TODO: get from config?
+        sub_element_class: ReferenceElement,
+        enumerator_class: self
+      )
+    end
+
+  end
+end