+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Access `.pantry` and `.clipboard` through an element instead of just its document (minor)
 * Add `suppress_solution` option to `BakeNumberedExercise` (minor)
 * Add strategy for American Government answer key (minor)
+* Add `BakeReferences` direction (minor)
 
 ## [3.2.0] - 2021-04-19